Refresh page stuck issue resolved

+++ src/app/_services/user-auth.service.ts
@@ -1,32 +1,46 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
+import { BrowserService } from './browser.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class UserAuthService {
-  constructor() { }
+  private _browserService = inject(BrowserService);
+  _isBrowser: boolean;
 
-  public setRoles(roles:[]) {
-    localStorage.setItem('roles', JSON.stringify(roles));
+  constructor() {
+    this._isBrowser = this._browserService.isBrowser();
+  }
+
+  public setRoles(roles: []) {
+    if (this._isBrowser)
+      localStorage.setItem('roles', JSON.stringify(roles));
   }
 
   public getRoles(): [] {
-    const item = localStorage.getItem('roles');
-    if(item)
-      return JSON.parse(item);
+    if (this._isBrowser) {
+      const item = localStorage.getItem('roles');
+      if (item)
+        return JSON.parse(item);
+      return [];
+    }
     return [];
   }
 
   public setToken(jwtToken: string) {
-    localStorage.setItem('jwtToken', jwtToken);
+    if (this._isBrowser)
+      localStorage.setItem('jwtToken', jwtToken);
   }
 
   public getToken() {
-    return localStorage.getItem('jwtToken');
+    if (this._isBrowser)
+      return localStorage.getItem('jwtToken');
+    return '';
   }
 
   public clearLocalStorage() {
-    localStorage.clear();
+    if (this._isBrowser)
+      localStorage.clear();
   }
 
   public isLoggedIn() {
@@ -34,20 +48,20 @@ export class UserAuthService {
   }
 
   public isAdmin() {
-    const roles : any[]= this.getRoles();
-    for(var i=0; i<roles.length; i++) {
-      if(roles[i]['roleName'] == 'Admin')
-      return true;
-   }
+    const roles: any[] = this.getRoles();
+    for (var i = 0; i < roles.length; i++) {
+      if (roles[i]['roleName'] == 'Admin')
+        return true;
+    }
     return false;
   }
 
   public isUser() {
-    const roles : any[]= this.getRoles();
-    for(var i=0; i<roles.length; i++) {
-    if(roles[i]['roleName'] == 'User')
-      return true;
-   }
+    const roles: any[] = this.getRoles();
+    for (var i = 0; i < roles.length; i++) {
+      if (roles[i]['roleName'] == 'User')
+        return true;
+    }
     return false;
   }
 }
